fix: fix message queueing
