+++ apps/nextjs/src/components/AppComponents/Chat/prompt-form.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef } from "react";
+import { useEffect, useRef } from "react";
 
 import type { UseChatHelpers } from "ai/react";
 
@@ -8,60 +8,32 @@ import {
   TooltipTrigger,
 } from "@/components/AppComponents/Chat/ui/tooltip";
 import { Icon } from "@/components/Icon";
-import { useLessonPlanTracking } from "@/lib/analytics/lessonPlanTrackingContext";
 import { useEnterSubmit } from "@/lib/hooks/use-enter-submit";
-import { useSidebar } from "@/lib/hooks/use-sidebar";
-import type { AilaStreamingStatus } from "@/stores/chatStore";
+import { useChatStore } from "@/stores/chatStore";
 
 export interface PromptFormProps
   extends Pick<UseChatHelpers, "input" | "setInput"> {
   onSubmit: (value: string) => void;
   hasMessages: boolean;
-  ailaStreamingStatus: AilaStreamingStatus;
-  queuedUserAction?: string | null;
-  queueUserAction?: (action: string) => void;
+  isDisabled: boolean;
 }
 
 export function PromptForm({
-  ailaStreamingStatus,
   onSubmit,
+  isDisabled,
   input,
   setInput,
   hasMessages,
-  queuedUserAction,
-  queueUserAction,
 }: Readonly<PromptFormProps>) {
   const { formRef, onKeyDown } = useEnterSubmit();
   const inputRef = useRef<HTMLTextAreaElement>(null);
-  const lessonPlanTracking = useLessonPlanTracking();
+  const queuedUserAction = useChatStore((state) => state.queuedUserAction);
 
   useEffect(() => {
     if (inputRef.current) {
       inputRef.current.focus();
     }
   }, []);
-
-  const sidebar = useSidebar();
-
-  const handleSubmit = useCallback(
-    (value: string) => {
-      setInput("");
-      if (sidebar.isSidebarOpen) {
-        sidebar.toggleSidebar();
-      }
-
-      lessonPlanTracking.onSubmitText(value);
-      if (queueUserAction) {
-        queueUserAction(value);
-      } else {
-        onSubmit(value);
-      }
-    },
-    [lessonPlanTracking, queueUserAction, onSubmit, setInput, sidebar],
-  );
-
-  const shouldAllowUserInput =
-    ["Idle", "Moderating"].includes(ailaStreamingStatus) && !queuedUserAction;
 
   return (
     <form
@@ -70,19 +42,19 @@ export function PromptForm({
         if (!input?.trim()) {
           return;
         }
-        handleSubmit(input);
+        onSubmit(input);
       }}
       ref={formRef}
     >
       <div
-        className={`${!shouldAllowUserInput ? "block" : "hidden"} h-[60px] w-full rounded-md border-2 border-oakGrey3 sm:hidden`}
+        className={`${isDisabled ? "block" : "hidden"} h-[60px] w-full rounded-md border-2 border-oakGrey3 sm:hidden`}
       />
       <div
-        className={`${!shouldAllowUserInput ? "hidden" : "flex"} relative max-h-60 w-full grow flex-col overflow-hidden rounded-md border-2 border-black bg-white pr-20 sm:flex`}
+        className={`${isDisabled ? "hidden" : "flex"} relative max-h-60 w-full grow flex-col overflow-hidden rounded-md border-2 border-black bg-white pr-20 sm:flex`}
       >
         <textarea
           data-testid="chat-input"
-          disabled={!shouldAllowUserInput}
+          disabled={isDisabled}
           ref={inputRef}
           tabIndex={0}
           onKeyDown={onKeyDown}
@@ -102,8 +74,8 @@ export function PromptForm({
               <button
                 data-testid="send-message"
                 type="submit"
-                className={`rounded-full bg-black p-4 ${!shouldAllowUserInput ? "cursor-not-allowed opacity-50" : ""}`}
-                disabled={!shouldAllowUserInput}
+                className={`rounded-full bg-black p-4 ${isDisabled ? "cursor-not-allowed opacity-50" : ""}`}
+                disabled={isDisabled}
               >
                 <Icon icon="chevron-right" color="white" size="sm" />
               </button>
